+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/activity/ActivityContext.kt
@@ -3,6 +3,7 @@ package io.github.pandier.intellijdiscordrp.activity
 import com.intellij.openapi.application.ApplicationInfo
 import com.intellij.openapi.application.ApplicationNamesInfo
 import com.intellij.openapi.application.ReadAction
+import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ProjectFileIndex
 import com.intellij.openapi.vfs.VfsUtil
@@ -22,6 +23,8 @@ class ActivityFileContext(
     val directoryName: String,
     val type: ActivityFileType,
     val typeName: String,
+    val line: Int?,
+    val lineCount: Int?,
 )
 
 class ActivityContext(
@@ -37,6 +40,7 @@ class ActivityContext(
         fun create(
             project: Project,
             file: VirtualFile? = null,
+            editor: Editor? = null,
             start: Instant = TimeTrackingService.getInstance(project).start,
         ): ActivityContext {
             val appInfo = ApplicationInfo.getInstance()
@@ -60,6 +64,8 @@ class ActivityContext(
                         directoryName = it.parent?.name ?: "",
                         type = activityFileType ?: ActivityFileType.OTHER,
                         typeName = activityFileType?.friendlyName ?: it.fileType.name,
+                        line = editor?.caretModel?.logicalPosition?.line?.plus(1),
+                        lineCount = editor?.document?.lineCount,
                     )
                 }
             )