+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/activity/ActivityDisplayMode.kt
@@ -72,7 +72,17 @@ enum class ActivityDisplayMode(
                 name = "file_dir_name",
                 description = "Name of the directory of the edited file",
                 getter = { file?.directoryName }
-            )
+            ),
+            ActivityVariable(
+                name = "file_line",
+                description = "Line number of the current line in the edited file",
+                getter = { file?.line?.toString() ?: "-" }
+            ),
+            ActivityVariable(
+                name = "file_line_count",
+                description = "Number of lines of the edited file",
+                getter = { file?.lineCount?.toString() ?: "-" }
+            ),
         ))
     );