+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/activity/icon/IconRepository.kt
@@ -13,7 +13,7 @@ package io.github.pandier.intellijdiscordrp.activity.icon
  * When these changes make it into the repository, the icons
  * will start displaying correctly.
  */
-private const val ICON_REPOSITORY_COMMIT = "d7bb2f716cc0f52487d1e3d92373913ea7bd7c86"
+private const val ICON_REPOSITORY_COMMIT = "4b741103e94585c469aefe0dbad6d525b7c9aace"
 const val ICON_REPOSITORY =
     "https://raw.githubusercontent.com/pandier/intellij-discord-rp/$ICON_REPOSITORY_COMMIT/icons"
 const val FILE_ICON_REPOSITORY = "$ICON_REPOSITORY/file"