+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/service/DiscordService.kt
@@ -3,6 +3,7 @@ package io.github.pandier.intellijdiscordrp.service
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
+import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.editor.EditorFactory
 import com.intellij.openapi.editor.ex.EditorEventMulticasterEx
 import com.intellij.openapi.project.Project
@@ -10,6 +11,8 @@ import com.intellij.openapi.vfs.VirtualFile
 import io.github.pandier.intellijdiscordrp.DiscordRichPresencePlugin
 import io.github.pandier.intellijdiscordrp.activity.ActivityContext
 import io.github.pandier.intellijdiscordrp.activity.currentActivityApplicationType
+import io.github.pandier.intellijdiscordrp.listener.RichPresenceCaretListener
+import io.github.pandier.intellijdiscordrp.listener.RichPresenceDocumentListener
 import io.github.pandier.intellijdiscordrp.listener.RichPresenceFocusChangeListener
 import io.github.pandier.intellijdiscordrp.settings.discordSettingsComponent
 import io.github.pandier.intellijdiscordrp.util.KPresenceLoggerAdapter
@@ -76,6 +79,8 @@ class DiscordService(
         val eventMulticaster = EditorFactory.getInstance().eventMulticaster
         val eventMulticasterEx = eventMulticaster as? EditorEventMulticasterEx
         eventMulticasterEx?.addFocusChangeListener(RichPresenceFocusChangeListener, this)
+        eventMulticasterEx?.addDocumentListener(RichPresenceDocumentListener, this)
+        eventMulticasterEx?.addCaretListener(RichPresenceCaretListener, this)
 
         // Connect to Discord client
         reconnectBackground()
@@ -174,19 +179,20 @@ class DiscordService(
      *
      * @see changeActivity
      */
-    fun changeActivityBackground(project: Project, file: VirtualFile?) {
+    fun changeActivityBackground(project: Project, file: VirtualFile?, editor: Editor?) {
         scope.launch(Dispatchers.Default) {
-            changeActivity(project, file)
+            changeActivity(project, file, editor)
         }
     }
 
     /**
      * Changes the activity to the given project and file.
+     * The [Editor] object is used for stuff like line count and caret position.
      *
      * @see changeActivity
      */
-    suspend fun changeActivity(project: Project, file: VirtualFile?) {
-        val activityContext = ActivityContext.create(project, file)
+    suspend fun changeActivity(project: Project, file: VirtualFile?, editor: Editor?) {
+        val activityContext = ActivityContext.create(project, file, editor)
         changeActivity(activityContext)
     }